requirements.txt
allure-behave==2.2.3b1
allure-python-commons==2.2.3b1
attrs==17.3.0
behave==1.2.5
enum34==1.1.6
parse==1.8.2
parse-type==0.4.2
pluggy==0.5.2
PyHamcrest==1.9.0
six==1.11.0
